add condition for appkey

+++ www/install/php/Update-22.10.0-beta.1.php
@@ -35,9 +35,10 @@ try {
 
     $errorMessage = "Unable to delete 'appKey' information from database";
     $pearDB->query("DELETE FROM `informations` WHERE `key` = 'appKey'");
-
-    $errorMessage = "Unable to drop 'app_key' from remote_servers table";
-    $pearDB->query("ALTER TABLE remote_servers DROP COLUMN `app_key`");
+    if ($pearDB->isColumnExist('remote_servers', 'app_key') === 1) {
+        $errorMessage = "Unable to drop 'app_key' from remote_servers table";
+        $pearDB->query("ALTER TABLE remote_servers DROP COLUMN `app_key`");
+    }
 
     $pearDB->commit();
 } catch (\Exception $e) {
